针对图片大于32767后超出区域绘制为空白的Bug
这个Bug现在将32767的限制放宽为65535，比这个尺寸更大的页面QT库在生成图片文件的时候会出错，故而暂时无法进一步增大。

+++ src/qt/src/gui/painting/qrasterdefs_p.h
@@ -867,9 +867,9 @@ QT_FT_BEGIN_HEADER
   /*                                                                       */
   typedef struct  QT_FT_Span_
   {
-    short           x;
+    int             x;
     unsigned short  len;
-    short y;
+    int             y;
     unsigned char   coverage;
   } QT_FT_Span;
